Add tests, refactor views into badgrsocialauth app
